refactoring of evaluators
