+++ bark_ml/evaluators/__init__.py
@@ -1,3 +1,2 @@
-from bark_ml.evaluators.goal_reached import GoalReached  # pylint: disable=unused-import
-from bark_ml.evaluators.reward_shaping import RewardShapingEvaluator  # pylint: disable=unused-import
 from bark_ml.evaluators.general_evaluator import GeneralEvaluator  # pylint: disable=unused-import
+from bark_ml.evaluators.evaluator_configs import *  # pylint: disable=unused-import
