--- bark_ml/evaluators/reward_shaping.py
@@ -1,21 +1,0 @@
-# Copyright (c) 2020 Patrick Hart, Julian Bernhard,
-# Klemens Esterle, Tobias Kessler
-#
-# This software is released under the MIT License.
-# https://opensource.org/licenses/MIT
-
-from bark_ml.evaluators.general_evaluator import *
-
-
-class RewardShapingEvaluator(GeneralEvaluator):
-  def __init__(self, params):
-    self._params = params["ML"]["RewardShapingEvaluator"]
-    super().__init__(
-      params=self._params,
-      bark_ml_eval_fns={
-        "collision_functor" : CollisionFunctor(self._params),
-        "drivable_area_functor" : DrivableAreaFunctor(self._params),
-        "step_count_functor" : StepCountFunctor(self._params),
-        "pot_center_functor": PotentialCenterlineFunctor(self._params),
-        "pot_vel_functor": PotentialVelocityFunctor(self._params)
-      })
